% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{print_pptx_pdf}
\alias{print_pptx_pdf}
\title{Export a .pptx file created using 'officer' package as a .pdf}
\usage{
print_pptx_pdf(x, target = NULL, ...)
}
\arguments{
\item{x}{document created using 'read_pptx()' function from 'officer' package}

\item{target}{path to the .pdf to be exported, default is a .pdf with the exported object's name in the current working directory}
}
\description{
This function exports the document as a temporary .pptx file, creates a
temporary .vbs file with the intended .pdf file name, runs the .vbs file in
command prompt, and then deletes the temporary files.
}
\examples{
ppt1 <- read_pptx()
ppt1 <- add_sheet(ppt1)
print_pptx_pdf(ppt1)
ppt2 <- read_pptx()
ppt2 <- add_sheet(ppt2)
print_pptx_pdf(ppt2, target = file.path(paste(getwd(),"/test.pdf",sep="")))
}
